requirements.txt
fastapi
uvicorn
pandas
python-dotenv
qdrant-client
google-generativeai
streamlit
